Add storage layer for ballots
