+++ backend/Cargo.toml
@@ -6,14 +6,6 @@ license = "MIT"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[[bin]]
-name = "grpc"
-path = "src/grpc.rs"
-
-[[bin]]
-name = "web"
-path = "src/web.rs"
-
 [dependencies]
 tonic = { workspace = true }
 prost = { workspace = true }
