+++ backend/src/bin/grpc.rs
@@ -1,21 +1,16 @@
-mod grpc_impl;
-mod model;
-mod settings;
-mod storage;
-
-use grpc_impl::close_poll::close_poll;
-use grpc_impl::create_poll::create_poll;
-use grpc_impl::get_poll_by_slug::get_poll_by_slug;
-use grpc_impl::submit_vote::submit_vote;
+use backend::grpc_impl::close_poll::close_poll;
+use backend::grpc_impl::create_poll::create_poll;
+use backend::grpc_impl::get_poll_by_slug::get_poll_by_slug;
+use backend::grpc_impl::submit_vote::submit_vote;
 use std::net::SocketAddr;
 
+use backend::settings::SETTINGS;
 use common::grpc::poll_service::poll_service_server::{PollService, PollServiceServer};
 use common::grpc::poll_service::{
     ClosePollRequest, ClosePollResponse, CreatePollRequest, CreatePollResponse,
     GetPollBySlugRequest, GetPollBySlugResponse, PollKind, PollKindsResponse, SubmitVoteRequest,
     SubmitVoteResponse,
 };
-use settings::SETTINGS;
 use tonic::codegen::http::Method;
 use tonic::{transport::Server, Request, Response, Status};
