+++ backend/src/storage/mod.rs
@@ -1,3 +1,5 @@
+pub mod ballot;
+
 use async_once::AsyncOnce;
 use lazy_static::lazy_static;
 use mongodb::{options::ClientOptions, Client, Database};
